Separated everything into components
