+++ shows/src/app/typings/IShow.ts
@@ -1,0 +1,6 @@
+interface IShow {
+  title: string;
+  description: string;
+  imageUrl: string | undefined;
+  averageRating: number | undefined;
+}
