+++ shows/src/app/typings/Review.ts
@@ -1,0 +1,4 @@
+interface review {
+  text: string;
+  rating: number | null;
+}
